Version Information for kafka-output unclear (#9032)
In the paragraph "compatibility" it is mentioned that versions 0.11 to 2.00 are supported.
In the description of the parameter "version" it is mentioned that all versions from 0.8.2.0 to 2.0.0 are valid.
It would be useful to mention the supported versions. This would avoid misinterpretations.
I assume that the information in the paragraph "compatibility" is correct.