Rename host.name to host.hostname and add config option for name (#9943)
Currently by default the hostname of the host on which the Beat is running is fed into host.name. Based on ECS this is now changed to `host.hostname`: https://github.com/elastic/ecs#-host-fields By default `host.name` stays the hostname of the host for backward compatibility. It is not set by the processor but by the pipeline.

If the config option `name` is changed in the Beat config, `host.name` will be set to this value as before. In addition a new config option was added to `add_host_metadata`:

```
processors:
  - add_host_metadata:
      name: foo
```

`name` for the host metadata processor can be set and will overwrite the current value of `host.name`.

It is also possible to set both config options as below:

```
name: bar

processors:
  - add_host_metadata:
      name: foo
```

The outcome event will contain the following two fields:

```
agent.name: bar
host.name: foo
```

Overall if nothing is configured, the behavior stays the same as before but allows additional options.