Mark indices with the beat name (#10313)
To facilitate APM re-indexing / migration scripts, it is very convenient to uniquely identify `apm` indices without searching for values in data.
Since everything is configurable by the user, right now we don't have any attribute that we can rely on. 

We have been recommended to put a mark on the `_meta` mapping. This enables us to set it in code without users tampering with it.