Add multiplier option to prometheus helper (#10994)
Prometheus histograms can contain buckets with keys in decimal
representation. They create events with dots in the field names, what
can be problematic when stored in Elasticsearch as they will create
sub-objects.

Add a new option to allow to store these keys in different units, so
precission is kept but in a different order of magnitude.