Remove `convert_timezone` option (#12469)
Timezone is now obtained from the system where filebeat is running, and
timezone conversion when parsing the timestamp is always done. This
timezone information is included in the `event.timezone` field, it can
be overriden with the `add_fields` processor, or removed with the
`remove_fields` processor.

Apply #12410 to the rest of modules.