Small pipeline enhancements (#12996)
Note: these changes are required to prepare for some refactoring in
filebeat.

Introduce OutputChooses publisher mode. Normally this is equivalent to
the default mode for most Beats. But we introduce a new mode here, as filebeat sets the default mode
to GuaranteedSend. This allows custom inputs to overwrite the default
mode in filebeat in the future.

Using beats.ProcessorList and the processors list to combine processors
from different configurations is a little tricky. To simplify the task,
we make sure that beat.ProcessorList implements processors.Processor.
We also export processors.NewList, that can be used to compose a set of
custom processors.Processor and beats.ProcessorList. The list returned
by processors.NewList also implements beats.ProcessorList.
It's not nice, but at least makes processors composable.