[Auditbeat] New system/socket dataset using kprobes tracing events (#13058)
This patch replaces the original `system/socket` dataset which used
netlink to periodically poll the system for open sockets.

In the new implementation, kprobe-based tracing is used to receive
events of interest related to function calls and their arguments
inside the kernel tcp/ip stack. From this stream of information the
dataset constructs a model of all the network sockets in use and the
processes running in the system at any given time.

This allows the dataset to produce flow-like events with information
like packets and bytes counters, originator process and user, with
little impact to system performance and without missing short-lived
sockets or processes.

The dataset is in beta.