Fail with error when autodiscover providers have no defined configs (#13078)
If one accidentally misconfigures an autodiscover provider and there are no defined configs no error is raised today. This patch adds an error in this situation and also does a better job adding context to other initialization errors.

You can test this with the following config which looks like it should work in heartbeat:

heartbeat.autodiscover:
  providers:
    - type: docker
      templates:
          condition:
            contains:
              docker.container.image: redis
          config:
            - type: tcp
              hosts: ["${data.host}:${data.port}"]
              schedule: "@every 1s"
              timeout: 1s

However, it does not because templates should be an array. The correct config is below:

heartbeat.autodiscover:
  providers:
    - type: docker
      templates:
        - condition:
            contains:
              docker.container.image: redis
          config:
            - type: tcp
              hosts: ["${data.host}:${data.port}"]
              schedule: "@every 1s"
              timeout: 1s