[Filebeat] Add dashboards to CEF module (#14342)
* Import Arcsight module dashboards from Logstash

Load dashboards to Kibana 8 via Logstash module setup.
Then use mage exportDashboard to download each of them to Beats.

* Change index pattern from arcsight-* to filebeat-*

* Make UUIDs unique from Logstash module

* Change [ArcSight] to [Filebeat CEF] in titles

* Fix dashboard IDs in navigation

* Update field names in dashboards

* Convert CEF extension value data types

Previously all of cef.extensions.* were string values. Now it uses the data types that are defined for each CEF field to do the translation within the processor. If the field is not the correct data type then the value is dropped from the event in order to avoid mapping exceptions at index time.

* Add fields for ArcSight to the mapping

These fields aren't specifically in CEF. They are commonly used in ArcSight.

* Exclude cef.extension fields from default_field and update dashboards

* Add descriptions to dashboards