Update fields.ecs.yml to ECS v1.5.0 (#16900)
* Update vendored elastic/ecs to v1.5.0

* Update fields.ecs.yml to v1.5.0

* Fix encoding issue in fields' docs

Fields' docs generation is failing in some systems when non-ascii
characters appear in the fields descriptions and/or examples.

* Patch fields.ecs.yml to avoid parser trouble

The fields docs generator was having trouble representing an example as
an object in a way that was consistent across different python3 versions
due to keys in a dict being serialised to string with different
orderings. This caused checks to fail.