Enable aarch64 (arm64) packaging by default (#17301)
This adds the aarch64 (arm64) architecture to the default list Linux package targets.

This will add three new artifacts to each beat project. For example:

- filebeat-$version-arm64.deb
- filebeat-$version-aarch64.rpm
- filebeat-$version-linux-arm64.tar.gz

I had to modify dockerlogbeat to honor the PLATFORMS selector. It was always trying
to build its linux/amd64 docker image even if that platform was not selected.