Remove billing from the list of AWS light metricsets (#21164)
On tests, loading any metricset from the AWS module is trying
to load the billing metricset as light metricset, what fails. This
shouldn't happen after #15011, but it is probably happening
because on tests, not all metricsets are registered.

billing metricset was refactored to a native implementation
recently, in #20527.

By now, remove billing from the list so tests can be executed.