Add UBI 8 image to the dependencies report (#21374)
Red Hat certification process requires to include the UBI image in the
list of dependencies for the project.
Append it to the current list of dependencies.
This requires to include a new sourceURL column in the CSV, because
UBI images have a different url and source url. This is kept the same
for existing dependencies.