Update filestream reader offset when line is skipped (#23417)
## What does this PR do?

This PR adds two previously missing offset updates to the `filestream` reader when a line is skipped.

## Why is it important?

The offset could be incorrect if Filebeat skips the line if the line should not be published because of user configuration in `export_line` or `import_line`

If the offset is not updated in the reader, the state information of newer published events become incorrect. This might lead to duplicated events if Filebeat is restarted.