Add support for recent PostgreSQL versions in Metricbeat (#24402)
Add testing with with versions 10, 11, 12 and 13.
Add fields available since version 10 in the activity
metricset to give more details about the backend and its waits.
Since version 12, `pg_stat_database` includes an entry for
shared resources that don't belong to a single database. Handle
this in tests and add docs and example document.
Handle metrics renamed in `pg_stat_statements` in version 13.
Regenerate data files with latest code and PostgreSQL 13.