Build Packetbeat packages serially (#25340)
Relates: https://github.com/elastic/beats/issues/24304

Try building Packetbeat packages serially as opposed to concurrent cross-compiles for various goos/goarch combos.
Test this to see if it impacts some of the spurious failures we see. I've never been able to reproduce the failure locally so
it's hard to know if this will fix it.