Report dataset metrics in monitoring log reporter (#25727)
This adds periodic reporting of "dataset" metrics to the logs (off by default). This can be enabled by setting `logging.metrics.namespaces: [stats, dataset]`. These are the same metrics available from the HTTP monitoring endpoint under the /dataset path. This primarily improves visibility of Filebeat and Metricbeat because they make use of the "dataset" namespace for input and metricset instance metrics.

Prior to 7.10 the Filebeat input metrics were reported in the logged metrics, but this was lost when the metrics moved to the "dataset" namespace.