Use Saved Objects API for exporting and importing dashboards (#27220)
## What does this PR do?

This PR adopts the Saved Objects API for exporting and importing dashboards and index templates. We have agreed to provide essential support, so there are a few limitations:

* You can only import assets they have exported using Beats 7.15 and the assets we ship with Beats.
* You can only export new assets, not legacy format.

## Why is it important?

We have to move forward to let Kibana remove the old API.

## How to test this PR locally

Try to setup the dashboards of a Beat:

```sh
./filebeat setup --dashboards
```

Export dashboards:

```sh
./filebeat export dashboard --id {my-id}
```
and:

```sh
go run dev-tools/cmd/dashboards/export_dashboards.go -dashboard {my-id}
```

## Notes to the reviewer

The changeset is huge, but most of it comes from converting the dashboards. I suggest you use the file filter and disable json, ndjson and yml files. The changes are moving json to ndjson format and changing the extensions in the module.yml files from json to ndjson.

Please also test it manually with both the setup command and `export_dashboards.go`.

## Related issues

Closes #20672